Port data to C
