+++ include/overlay_76.h
@@ -1,0 +1,6 @@
+typedef struct {
+    u8 unk0;
+    u8 unk1;
+    u8 unk2;
+    u8 unk3;
+} UnkOv021E60F6;
