+++ main.lsf
@@ -886,8 +886,9 @@ Overlay OVY_75
 Overlay OVY_76
 {
     After main
-    Object src/overlay_76_src.o
-    Object asm/overlay_76.o
+    Object src/overlay_76_data.o
+    Object src/overlay_76.o
+    Object src/overlay_76_data2.o
 }
 Overlay OVY_77
 {
