+++ src/overlay_76_data.c
@@ -1,0 +1,8 @@
+#include "overlay_76.h"
+
+const UnkOv021E60F6 ov76_021E6E88[4] = {
+    { 0x49, 0x00, 0x00, 0x00 },
+    { 0x49, 0x00, 0x00, 0x00 },
+    { 0x49, 0x00, 0x00, 0x00 },
+    { 0x07, 0x07, 0x07, 0x07 },
+};
