+++ src/overlay_76_data2.c
@@ -1,0 +1,1 @@
+const u8 ov76_021E878C[8] = { 6, 1, 2, 4, 3, 5, 0, 0 };
